requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
librosa>=0.10.0
numpy>=1.24.0
scipy>=1.10.0
pretty_midi>=0.2.10
soundfile>=0.12.1
transformers>=4.30.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
tqdm>=4.65.0
hydra-core>=1.3.0
wandb>=0.15.0  # For experiment tracking 